Move actor into system
