+++ radix-engine/src/kernel/mod.rs
@@ -1,4 +1,3 @@
-pub mod actor;
 pub mod call_frame;
 pub mod heap;
 pub mod id_allocator;
