+++ radix-engine/src/system/mod.rs
@@ -14,3 +14,4 @@ pub mod system_modules;
 pub mod system_substates;
 pub mod system_type_checker;
 pub mod type_info;
+pub mod actor;
