+++ radix-engine/src/system/system_callback.rs
@@ -5,10 +5,10 @@ use crate::blueprints::identity::IDENTITY_CREATE_VIRTUAL_ED25519_ID;
 use crate::blueprints::identity::IDENTITY_CREATE_VIRTUAL_SECP256K1_ID;
 use crate::blueprints::transaction_processor::TransactionProcessorRunInputEfficientEncodable;
 use crate::errors::*;
-use crate::kernel::actor::Actor;
-use crate::kernel::actor::BlueprintHookActor;
-use crate::kernel::actor::FunctionActor;
-use crate::kernel::actor::MethodActor;
+use crate::system::actor::Actor;
+use crate::system::actor::BlueprintHookActor;
+use crate::system::actor::FunctionActor;
+use crate::system::actor::MethodActor;
 use crate::kernel::call_frame::CallFrameMessage;
 use crate::kernel::kernel_api::{KernelApi, KernelInvocation};
 use crate::kernel::kernel_api::{KernelInternalApi, KernelSubstateApi};
