+++ radix-engine/src/system/system_modules/costing/costing_entry.rs
@@ -1,5 +1,5 @@
 use super::FeeTable;
-use crate::kernel::actor::Actor;
+use crate::system::actor::Actor;
 use crate::kernel::kernel_callback_api::{
     CloseSubstateEvent, CreateNodeEvent, DrainSubstatesEvent, DropNodeEvent, MoveModuleEvent,
     OpenSubstateEvent, ReadSubstateEvent, RemoveSubstateEvent, ScanKeysEvent,
